Decouple cgal primitives into primitives-cgal.cc
